+++ changelog.md
@@ -31,6 +31,7 @@
 * [3975](https://github.com/zeta-chain/node/pull/3975) - add v2 connector deployment to the e2e test.
 * [3976](https://github.com/zeta-chain/node/pull/3976) - add connector fund migration e2e test using contracts only.
 * [3999](https://github.com/zeta-chain/node/pull/3999) - run simulation tests nightly
+* [3985](https://github.com/zeta-chain/node/pull/3985)- add e2e tests for deposit and withdraw with big payload
 
 ## v31.0.0